first proper test
