--- tests/test_one.py
@@ -1,6 +1,0 @@
-import math
-
-def test_sqrt():
-   num = 25
-   assert math.sqrt(num) == 5
-
